requirements.txt
gradio>=4.0.0
requests>=2.28.0
modal>=0.55.0
pyttsx3>=2.90
diffusers>=0.24.0
transformers>=4.34.0
torch>=2.0.0
accelerate>=0.23.0
imageio>=2.31.0
Pillow>=10.0.0
